+++ src/behaviors/miner.lua
@@ -39,8 +39,8 @@ local tick = 1
 ---   @field timeToMine number (default: `15`) seconds it takes the miner to mine an asteroid
 ---   @field timeToGoHome number (default: `900`) seconds the miner tries to mine asteroids before giving up and returning home disappointed
 ---   @field mineDistance number (default: `beamWeaponRange`) units how close the miner needs to be to the asteroid
----   @field maxDistanceFromHome number (default: `getLongRangeRadarRange`) units how far away from home the miner looks for asteroids
----   @field maxDistanceToNext number (default: `getLongRangeRadarRange / 2`) units how far from the current asteroid the miner will look for a next one
+---   @field maxDistanceFromHome number (default: `30000`) units how far away from home the miner looks for asteroids
+---   @field maxDistanceToNext number (default: `15000`) units how far from the current asteroid the miner will look for a next one
 ---   @field onHeadingAsteroid function gets `miner` and `asteroid` when the miner is flying towards an asteroid
 ---   @field onAsteroidMined function gets `miner`, `asteroid` and the return of `whenMined`
 ---   @field onHeadingHome function gets `miner`, `asteroid` and all the gathered products
@@ -81,9 +81,9 @@ Ship.behaveAsMiner = function (self, ship, homeStation, whenMined, config)
     if not isNumber(config.timeToGoHome) then error("Expected timeToGoHome to be a number, but got " .. typeInspect(config.timeToGoHome), 2) end
     config.mineDistance = config.mineDistance or ship:getBeamWeaponRange(0)
     if not isNumber(config.mineDistance) then error("Expected mineDistance to be a number, but got " .. typeInspect(config.mineDistance), 2) end
-    config.maxDistanceFromHome = config.maxDistanceFromHome or getLongRangeRadarRange()
+    config.maxDistanceFromHome = config.maxDistanceFromHome or 30000
     if not isNumber(config.maxDistanceFromHome) then error("Expected maxDistanceFromHome to be a number, but got " .. typeInspect(config.maxDistanceFromHome), 2) end
-    config.maxDistanceToNext = config.maxDistanceToNext or (getLongRangeRadarRange() / 2)
+    config.maxDistanceToNext = config.maxDistanceToNext or 15000
     if not isNumber(config.maxDistanceToNext) then error("Expected maxDistanceToNext to be a number, but got " .. typeInspect(config.maxDistanceToNext), 2) end
     config.onHeadingAsteroid = config.onHeadingAsteroid or function() end
     if not isFunction(config.onHeadingAsteroid) then error("Expected onHeadingAsteroid to be a function, but got " .. typeInspect(config.onHeadingAsteroid), 2) end